Moved exclusive files-related configurations keys to IOConfigurationTrait
