+++ lib/Phpfastcache/Config/IOConfigurationOptionTrait.php
@@ -14,6 +14,10 @@
 declare(strict_types=1);
 namespace Phpfastcache\Config;
 
+use Phpfastcache\Exceptions\PhpfastcacheInvalidConfigurationException;
+
+const SAFE_FILE_EXTENSIONS = 'txt|cache|db|pfc';
+
 trait IOConfigurationOptionTrait
 {
     /**
@@ -32,6 +36,11 @@ trait IOConfigurationOptionTrait
     protected $securityKey = '';
 
     /**
+     * @var string
+     */
+    protected $cacheFileExtension = 'txt';
+
+    /**
      * @return string
      */
     public function getSecurityKey(): string
@@ -43,7 +52,7 @@ trait IOConfigurationOptionTrait
      * @param string $securityKey
      * @return Config
      */
-    public function setSecurityKey(string $securityKey): ConfigurationOptionInterface
+    public function setSecurityKey(string $securityKey): self
     {
         $this->securityKey = $securityKey;
 
@@ -81,9 +90,44 @@ trait IOConfigurationOptionTrait
      * @param bool $secureFileManipulation
      * @return self
      */
-    public function setSecureFileManipulation(bool $secureFileManipulation): ConfigurationOptionInterface
+    public function setSecureFileManipulation(bool $secureFileManipulation): self
     {
         $this->secureFileManipulation = $secureFileManipulation;
+        return $this;
+    }
+
+
+    /**
+     * @return string
+     */
+    public function getCacheFileExtension(): string
+    {
+        return $this->cacheFileExtension;
+    }
+
+    /**
+     * @param string $cacheFileExtension
+     * @return self
+     * @throws PhpfastcacheInvalidConfigurationException
+     */
+    public function setCacheFileExtension(string $cacheFileExtension): self
+    {
+        /**
+         * Feel free to propose your own one
+         * by opening a pull request :)
+         */
+        $safeFileExtensions = \explode('|', SAFE_FILE_EXTENSIONS);
+
+        if (\strpos($cacheFileExtension, '.') !== false) {
+            throw new PhpfastcacheInvalidConfigurationException('cacheFileExtension cannot contain a dot "."');
+        }
+        if (!\in_array($cacheFileExtension, $safeFileExtensions, true)) {
+            throw new PhpfastcacheInvalidConfigurationException(
+                "Extension \"{$cacheFileExtension}\" is not safe, currently allowed extension names: " . \implode(', ', $safeFileExtensions)
+            );
+        }
+
+        $this->cacheFileExtension = $cacheFileExtension;
         return $this;
     }
 }
