+++ tests/issues/Github-467.test.php
@@ -6,31 +6,31 @@
  */
 
 use Phpfastcache\CacheManager;
-use Phpfastcache\Config\ConfigurationOption;
+use Phpfastcache\Drivers\Files\Config as FilesConfig;
 use Phpfastcache\Exceptions\PhpfastcacheInvalidConfigurationException;
 use Phpfastcache\Helper\TestHelper;
 
 chdir(__DIR__);
 require_once __DIR__ . '/../../vendor/autoload.php';
 $testHelper = new TestHelper('Github issue #467 - Allow to specify the file extension in the File Driver');
-CacheManager::setDefaultConfig(new ConfigurationOption(['path' => __DIR__ . '/../../cache']));
+CacheManager::setDefaultConfig(new FilesConfig(['path' => __DIR__ . '/../../cache']));
 
 try{
-    $cacheInstance = CacheManager::getInstance('Files', new ConfigurationOption(['cacheFileExtension' => 'php']));
+    $cacheInstance = CacheManager::getInstance('Files', new FilesConfig(['cacheFileExtension' => 'php']));
     $testHelper->printFailText('No error thrown while trying to setup a dangerous file extension');
 }catch(PhpfastcacheInvalidConfigurationException $e){
     $testHelper->printPassText('An error has been thrown while trying to setup a dangerous file extension');
 }
 
 try{
-    $cacheInstance = CacheManager::getInstance('Files', new ConfigurationOption(['cacheFileExtension' => '.cache']));
+    $cacheInstance = CacheManager::getInstance('Files', new FilesConfig(['cacheFileExtension' => '.cache']));
     $testHelper->printFailText('No error thrown while trying to setup a dotted file extension');
 }catch(PhpfastcacheInvalidConfigurationException $e){
     $testHelper->printPassText('An error has been thrown while trying to setup a dotted file extension');
 }
 
 try{
-    $cacheInstance = CacheManager::getInstance('Files', new ConfigurationOption(['cacheFileExtension' => 'cache']));
+    $cacheInstance = CacheManager::getInstance('Files', new FilesConfig(['cacheFileExtension' => 'cache']));
     $testHelper->printPassText('No error thrown while trying to setup a safe file extension');
 }catch(PhpfastcacheInvalidConfigurationException $e){
     $testHelper->printFailText('An error has been thrown while trying to setup a safe file extension');
